Refactor unstable MSC4108 implementation to prepare for updated version
